% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse.ACS.blank.R
\name{analyse.ACs.blank}
\alias{analyse.ACs.blank}
\title{Analyse AC-s blank (Pure Water) data}
\usage{
analyse.ACs.blank(Twater = 19, Tref = 20.3, AC = TRUE)
}
\arguments{
\item{Twater}{is the pure water temperature during the calibration
experiment on the field}

\item{Tref}{is the pure water temperature reported by Wet Labs during
the ACs calibration.}

\item{AC}{is a logical parameter indicating whether or not the a-beam
and c-beam were calibrated at the same time on the field. If TRUE, it will
assumes that both beams were calibrated at the same time and the user
will be prompted to select only one file. If FALSE, each beam was calibrated
separately and the user will be prompted to select one file for a-beam calibration
and one file for c-beam calibration.}
}
\value{
It returns a list with a.mean, a.smooth, a.sd,
c.mean, c.smooth, c.sd, c.wl, a.wl
}
\description{
Analyse AC-s blank (Pure Water) data
}
\details{
The user will be prompted to select the files and to click on a plot to
select a portion of the file (start and end) where a or c values were stable
and minimum. The program will average and smooth the spectral a and c values
for the selected portion and stored them in a RData file (i.e. ACS.blank list).

The RData file could be used in the data processing for ACs blank correction
in \code{\link{correct.merge.IOP.profile}}. The path of the file need to be
included in the cal.info.dat (see \code{\link{correct.merge.IOP.profile}})
}
\author{
Simon Belanger
}
