% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.ASPH.Vin.R
\name{check.ASPH.Vin}
\alias{check.ASPH.Vin}
\title{Check A-sphere batteries voltage input (Vin)}
\usage{
check.ASPH.Vin(filen)
}
\arguments{
\item{filen}{is the file name of the ASPH file.}
}
\value{
A plot showing Vin as a function of time.
}
\description{
It produces a diagnostic plot to evaluate if the
batteries need to be changed.
}
\details{
Note that only the a-sphere file processed using the IGOR
template version "a-Sphere Processing Template 201.pxt"
or higher (>201) include the Vin
}
\author{
Simon Bélanger
}

