% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.aTOT.for.COPS.R
\name{compute.aTOT.for.COPS}
\alias{compute.aTOT.for.COPS}
\title{Compute the total absorption for COPS processing of LuZ profile}
\usage{
compute.aTOT.for.COPS(pathIOP = ".", cast = "down",
  depth.interval = c(0.75, 2.1), instrument = "ASPH")
}
\arguments{
\item{pathIOP}{is the path where IOPs data are located
(i.e. IOPs.fitted.down.RData or IOPs.fitted.up.RData). These files
are generated by \code{\link{correct.merge.IOP.profile}}.
IMPORTANT NOTE: a COPS folder with files preliminarly processed using
the \pkg{Cops} package must be present in the parent folder "YYYYMMDD_StationXXX".
The \pkg{Cops} processing creates a file named absorption.cops.dat
that will be automatically edited.}

\item{cast}{is either "down" or "up" indicating whether the down cast
or the up cast will be used to average the non-water absorption.
The default is "down".}

\item{depth.interval}{is the depth interval for which the non-water
absorption will be avaraged.
The default is c(0.75,2.1)}

\item{a.instrument}{indicates whether the non-water absorption will be
taken from the a-sphere (ASPH) or the ac-s (ACS).
The default is "ASPH".}
}
\value{
It returns a vector of total absorption.
}
\description{
Compute the total absorption for COPS instrument self-shadow
correction of LuZ. It is run only after a preliminary processing
of the COPS data using the \pkg{Cops} package.
}
\details{
The program will open a RData file located in ../COPS/BIN folder to get
the wavelengths available on the COPS. It will edit the file
../COPS/absorption.cops.dat and produce a PNG figure showing the spectral
absorption that will be used in the COPS processing for instrument self-shadow
correction.

The pure water absorption is taken from a composite table
(Kou etal 1993, Pope and Fry 1997, etc).
}
\author{
Simon Bélanger
}
\seealso{
\code{\link{correct.merge.IOP.profile}} and
\code{\link{IOPs.go}}
}

