% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.bb.discrete.from.ECO.R
\name{compute.bb.discrete.from.ECO}
\alias{compute.bb.discrete.from.ECO}
\title{Backscattering processing from ECO sensor at discrete depth}
\usage{
compute.bb.discrete.from.ECO(eco, ECO.type = "VFS3", ECO.bands = NA,
  Station = NA, DateTime, Depth = NA, Anw = NA, Salinity = 0,
  start = 999, end = 999)
}
\arguments{
\item{eco}{is a list containing calibrated VSF data returned by 
\code{\link{apply.ECO.cal}}}

\item{ECO.type}{is a character string for the type of ECO meter: BB9, BB3, VSF3.
(Default is VSF3)}

\item{ECO.bands}{must be provided if ECO.type == VSF3. It is a character string indicating the bands (BLUE, GREEN and RED), 
i.e. either "B", "G", "R" or "BG", or "BGR".}

\item{Station}{is the station ID}

\item{DateTime}{is a Date and Time stamp in POSIXct format}

\item{Depth}{of sampling if data collected in discrete mode}

\item{Anw}{Non-water absorption for the corresponding bands of the ECO meter.}

\item{Salinity}{Salinity (Default = 0)}

\item{start}{is the begining of the cast in seconds after the instrument warming period.
If 999 then the user is prompt to click on the plot of 
VSF versus Time to choose the start and the end of the cast interactively.
(Default = 1)}

\item{end}{is the end of the cast in second. If 999 then it takes
the end of the cast. (Default = 999)}
}
\description{
Backscattering processing from ECO sensor at discrete depth
}
\author{
Simon Belanger
}
