% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.report.R
\name{create.report}
\alias{create.report}
\title{Create a PDF report for a vertical profile performed using an optical package}
\usage{
create.report(dirdat)
}
\arguments{
\item{dirdat}{is the current directory path contaning the RData file to plot.}
}
\description{
A PDF report is generated using knitr package facilities and a *.Rnw
template. It provides an efficent mean to visualise the data.
It can be called from the higher level function \code{\link{IOPs.go}}, or
run in the command line as along as IOPs.RData and IOPs.fitted.down.RData have been
generated by \code{\link{correct.merge.IOP.profile}}.
}
\details{
The program edit a template named IOP_Station_TEMPLATE.Rnw located in the package data folder.
It will extract a number of informations from the current
data directory and RData files, cast.info.dat and instruments.dat files. For example,
the Latitude/Longitude are extracted from the cast.info.dat file to
create a map (Figure 1 of the report).
Next it uses the brew package to update the Rnw files.

The Rnw files are converted into PDF using knitr package. Both files will
be saved in dirdat and a "figures" folder is created with
a PDF file for each figures found in the report.
}
\author{
Simon Bélanger
}
\seealso{
\code{\link{correct.merge.IOP.profile}}, \code{\link{IOPs.go}}
}

