% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.functions.R
\name{fit.with.loess}
\alias{fit.with.loess}
\title{A modified version of the loess adapted to spectral data.}
\usage{
fit.with.loess(waves, Depth, aop, span, depth.fitted,
  span.wave.correction = FALSE, verbose = FALSE)
}
\description{
A modified version of the loess adapted to spectral data.
}
