% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.ECOVSF.R
\name{process.ECOVSF}
\alias{process.ECOVSF}
\title{ECO-VSF processing for discrete depth (or profile sampling not implemented yet)}
\usage{
process.ECOVSF(eco, station = NA, DateTime, Depth = NA, B.Anw = NA,
  G.Anw = NA, Salinity = 0, start = 1, end = 999)
}
\arguments{
\item{eco}{is a list containing calibrated VSF data returned by 
\code{\link{apply.ECOVSF.cal}}}

\item{station}{is the station ID}

\item{DateTime}{is a Date and Time stamp in POSIXct format}

\item{Depth}{of sampling if data collected in discrete mode}

\item{B.Anw}{Non-water absorption for the 470 nm band
(Default = NA; no correction applied)}

\item{G.Anw}{Non-water absorption for the 532 nm band
(Default = NA; no correction applied)}

\item{Salinity}{Salinity (Default = 0)}

\item{start}{is the begining of the cast in seconds. 
If 999 then the user is prompt to click on the plot of 
VSF versus Time to choose the start and the end of the cast interactively.
(Default = 1)}

\item{end}{is the end of the cast in second. If 999 then it takes
the end of the cast. (Default = 999)}
}
\description{
ECO-VSF processing for discrete depth (or profile sampling not implemented yet)
}
\author{
Simon Belanger
}
