% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.IOPs.R
\name{process.IOPs}
\alias{process.IOPs}
\title{Reads the parameters and instrument files and launch processing}
\usage{
process.IOPs(dirdat)
}
\arguments{
\item{dirdat}{is the directory where the instrument files are stored.}
}
\description{
It call the reading functions for cast.info.dat and cal.info.dat,
instruments.dat. If they are not present in the directory, it
creates a copy of the templates proveded with the package.
Finaly it calls \code{\link{correct.merge.IOP.profile}}.
}
\seealso{
\code{\link{correct.merge.IOP.profile}} and
\code{\link{IOPs.go}}
}
\author{
Simon B\elanger
}
