% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ACs.R
\name{read.ACs}
\alias{read.ACs}
\title{Read an AC-s file}
\usage{
read.ACs(filen)
}
\arguments{
\item{filen}{is an ACs file name}
}
\value{
A list with Timer, c.wl, a.wl, c, a,
iTemp, Sal, Depth, XTemp, Cond
}
\description{
Read an AC-s file
}
\author{
Simon Belanger
}
