% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ASPH.R
\name{read.ASPH}
\alias{read.ASPH}
\title{Read A-Sphere ASCII file as exported by IGOR software.}
\usage{
read.ASPH(filen, skip = 0)
}
\arguments{
\item{filen}{is the ASPH file name}

\item{skip}{is the number of lines to skip at the begining of the file.
This was added to deal with files containing several casts.

 @return A list with time, depth, Vin and wl vector and a matrix
  a containing the spectral absorption.}
}
\description{
Read A-Sphere ASCII file as exported by IGOR software.
}
\author{
Simon Bélanger
}

