% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.BB3.R
\name{read.BB3}
\alias{read.BB3}
\title{Read s BB-3 file}
\usage{
read.BB3(filen)
}
\arguments{
\item{filen}{is a BB-3 file name}
}
\value{
A list with Timer, Beta, BetaP, bbP, bb, wl
}
\description{
Read s BB-3 file
}
\author{
Simon Bélanger
}
