% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.CTD.DH4.R
\name{read.CTD.DH4}
\alias{read.CTD.DH4}
\title{Read a CTD (MicroCAT) file}
\usage{
read.CTD.DH4(filen)
}
\arguments{
\item{filen}{is a CTD file name}
}
\value{
A list with Timer, Temp, Cond, Depth, Sal
}
\description{
Read a CTD file in ASCII format as recorded
in the DH4
}
\author{
Simon Bélanger
}

