% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.CTD.R
\name{read.CTD}
\alias{read.CTD}
\title{Read a CTD SBE19+ file}
\usage{
read.CTD(filen)
}
\arguments{
\item{filen}{is a CTD file name}
}
\value{
A list with Time, Temp, Depth, Sal,
Density, Voltage, Tpot
}
\description{
Read a CTD SBE19+ file in ASCII format as recorded in the MiniDAS
}
\author{
Simon Bélanger
}

