% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ECO.dev.file.R
\name{read.ECO.dev.file}
\alias{read.ECO.dev.file}
\title{Read a WetLabs Device file for ECO meter (VSF3, BB9, BB3)}
\usage{
read.ECO.dev.file(filen, ECO.type = "VSF3")
}
\arguments{
\item{filen}{is device file name}

\item{ECO.type}{is a character string for the type of ECO meter: BB9, BB3, VSF3.
(Default is VSF3)}
}
\value{
A list including a dataframe for the calibration coefficients, 
the ECO serial number and the data of the calibration.
}
\description{
Read a WetLabs Device file for ECO meter (VSF3, BB9, BB3)
}
\author{
Simon Bélanger
}
