% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ECOVSF.ISMER.R
\name{read.ECOVSF.ISMER}
\alias{read.ECOVSF.ISMER}
\title{Read the ISMER ECO-VSF data}
\usage{
read.ECOVSF.ISMER(eco.file)
}
\arguments{
\item{eco.file}{is the file name of the ECO-VSF data}
}
\value{
A dataframe with the calibrated data
}
\description{
Read the ISMER ECO-VSF data
}
\author{
Simon Bélanger
}
