% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.FLBBCD.R
\name{read.FLBBCD}
\alias{read.FLBBCD}
\title{Read an ECO triplet for chlrophyll, bb700 and cdom (FLBBCD) file}
\usage{
read.FLBBCD(filen)
}
\arguments{
\item{filen}{is a FLBBCD file name}
}
\value{
A list with Timer, Beta700, BetaP700,
bbP700, bb700, FCHL, FDOM
}
\description{
Read an ECO triplet for chlrophyll, bb700 and cdom (FLBBCD) file
}
\author{
Simon Bélanger
}

