% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.FLCHL.R
\name{read.FLCHL}
\alias{read.FLCHL}
\title{Read a chlorophyll fluorescence (FLCHL) file}
\usage{
read.FLCHL(filen)
}
\arguments{
\item{filen}{is a FLCHL file name}
}
\value{
A list with Timer and FCHL
}
\description{
Read a chlorophyll fluorescence (FLCHL) file
}
\author{
Simon Belanger
}
