% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.FLECO.R
\name{read.FLECO}
\alias{read.FLECO}
\title{Read an ECO triplet for CDOM fluorescence (FLECO) file}
\usage{
read.FLECO(filen)
}
\arguments{
\item{filen}{is a FLECO file name}
}
\value{
A list with FL, ex, em, TempIntrument
}
\description{
Read an ECO triplet for CDOM fluorescence (FLECO) file
}
\author{
Simon Bélanger
}

