% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.HS6.R
\name{read.HS6}
\alias{read.HS6}
\title{Read s Hydroscat-6 file}
\usage{
read.HS6(filen)
}
\arguments{
\item{filen}{is a Hydroscat-6 file name}
}
\value{
A list with Time, depth, wl,bb, bbu,beta,betau,
fluo,fluou,betafl
}
\description{
Read s Hydroscat-6 file
}
\author{
Simon Bélanger
}
