% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.LISST.R
\name{read.LISST}
\alias{read.LISST}
\title{Read s LISST file in format *.asc}
\usage{
read.LISST(filen)
}
\arguments{
\item{filen}{is a LISST file name}
}
\value{
A list with time, Depth, PSD, lower_size_bins, upper_size_bins,
median_bins,temperature, transsmision, c670
}
\description{
Read s LISST file in format *.asc
}
\author{
Simon Belanger
}
