% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.vsf3.dev.file.R
\name{read.vsf3.dev.file}
\alias{read.vsf3.dev.file}
\title{Read a WetLabs Device file for ECO-VSF3}
\usage{
read.vsf3.dev.file(filen)
}
\arguments{
\item{filen}{is DEV file name}
}
\value{
A dataframe
}
\description{
Read a WetLabs Device file for ECO-VSF3
}
\author{
Simon Belanger
}
