% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.LISST100.R
\name{reformat.LISST100}
\alias{reformat.LISST100}
\title{Convert raw LISST ASCII to raw binary file format (*.DAT)}
\usage{
reformat.LISST100(filen)
}
\arguments{
\item{filen}{is a LISST 100 file name}
}
\value{
It returns the matrix with 40 columns that have been
converted into binary.
}
\description{
Convert raw LISST ASCII, when connected to a DH4 for example, to a binary file format
readable by LISST-SOP software for processing
}
\details{
When operate using a DH4, the LISST output is stored as ASCII
strings that are unreadable by the LISST-SOP software for further
processing. This program reformat the ASCII output a creates a
binary file (*.DAT) readable by LISST-SOP software, which can be use to
calculate the Particles Size Distribution (PSD) and output a file
*.asc. The *.asc file can be read by \code{\link{read.LISST}}.
}
\author{
Simon Belanger
}
