% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.process.ECO.batch.R
\name{run.process.ECO.batch}
\alias{run.process.ECO.batch}
\title{Run batch processing to convert ECO meter data in raw format to backscattering coefficients.}
\usage{
run.process.ECO.batch(log.file = "log.txt", data.path = "./")
}
\arguments{
\item{log.file}{is the name of the ASCII file (semicolumn-delimated text) containing the
list of samples to process (see details below).}

\item{data.path}{is the full path where the ECO data are stored}
}
\description{
The processing includes the application of the calibration coefficients to convert the raw numerical counts into VSF. 
 The dark offsets may be taken from the calibration file (device file) or from the dark measurements taken on the field. 
 Next, if absorption coefficients are provided, the VSF is corrected for loss of photons 
 due to attenuation along the pathlength. 
 The VSF is finaly converted in to total backscattering and particles backscattering.
}
\details{
The most important thing to do before runing this programm is to prepare
the log.file. This file contains fields (semi-column delimiter) : 
ECO.filename; dev.file; ECO.type; ECO.bands; Station; Date; UTCTime; 
Depth; Salinity; dark.file; start; end; process; Anw1;..;AnwX
The ASCII files are:
\itemize{
 \item{ECO.filename: }{is the file name of an ECO file without the path. It must be put in a sub-folder named raw/ in the data.path;}
 \item{dev.file: }{is the full path and name of the device file (i.e. the calibration file) of the actual instrument;}
 \item{ECO.type: }{is a character string for the type of ECO meter: BB9, BB3, VSF3.}
 \item{ECO.bands: }{is a character string indicating the bands of the VSF3 meter. 
 It could take one of these character: "B", "G", "R" or "BG", or "BGR". 
 It is ignore when the ECO type is a BB meter}
 \item{Station: }{is the station ID. It will be store in the RData output;}
 \item{Date: }{is the date in format YYYYMMDD}
 \item{UTCTime: }{is the is the time and minute in HH:MM format}
 \item{Salinity: }{is the salinity of the sample.}
 \item{dark.file: }{is the name of the file containing raw dark measurements made 
 on the field using black tape places on the LED. 
 It must be put in a sub-folder named dark/ in the data.path and be consistent with the ECO.type}
 \item{start: }{is the begining of the cast in seconds after the instrument warming period.  
If 999 then the user is prompt to click on the plot of 
VSF versus Time to choose the start and the end of the cast interactively.}
 \item{end: }{is the end of the cast in second. If 999 then it takes
the end of the cast.}
 \item{process: }{is a boolean value indicating if the file has to be process (1) or not (0)  }
 \item{Anw1..AnwX: }{is the non-water absorption for each wavelength of the ECO meter. 
 For a BB9, for example, nine columns in expected}
 }
 
 The code will process each file one by one and will output a RData file for each raw file processed. 
 
 @author  Simon Belanger
 @export
}
