% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.process.ECOVSF.batch.R
\name{run.process.ECOVSF.batch}
\alias{run.process.ECOVSF.batch}
\title{Run batch processing for ECO-VSF}
\usage{
run.process.ECOVSF.batch(log.file = "log.txt", data.path = "./",
  dev.file = "VSF30025.DEV")
}
\arguments{
\item{log.file}{is the name of the ASCII file (CSV or space-delimated text) containing the
list of samples to process (see details below).}

\item{data.path}{is the full path where the ECO data are stored}

\item{dev.file}{is the Wetlabs device file}
}
\description{
Run batch processing for ECO-VSF
}
\details{
The most important thing to do before runing this programm is to prepare
the log.file. This file contains 11 fields : 
Eco.filename Station Date Depth B.Anw G.Anw Salinity dark.file start end process
}
