% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.aph.bricaud.R
\name{spectral.aph.bricaud}
\alias{spectral.aph.bricaud}
\title{Compute spectral phyotplankton absorption
using Bricaud et al (1995, 1998) statistics
for case 1 waters}
\usage{
spectral.aph.bricaud(wl, chl)
}
\description{
Compute spectral phyotplankton absorption
using Bricaud et al (1995, 1998) statistics
for case 1 waters
}

