% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.aw.R
\name{spectral.aw}
\alias{spectral.aw}
\title{Pure water absorption spectrum}
\usage{
spectral.aw(waves, MOREL = FALSE)
}
\arguments{
\item{waves}{is the wavelenght (could be a vector)}

\item{MOREL}{is a logical to compute pure water absorption from Morel 1988 table.}
}
\description{
Compute spectral absorption coefficient for pure water from 280 to 900 nm
from Pope and Fry 1997 and Buiteveld 1994 experiments.
Optionnaly, K_w from Morel 1988 minus b_bw can be requested
}
\author{
Simon Belanger
}
