% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.bw.R
\name{spectral.bw}
\alias{spectral.bw}
\title{Pure water scattering coefficient spectrum}
\usage{
spectral.bw(waves)
}
\arguments{
\item{waves}{is the wavelenght (could be a vector)}
}
\description{
Compute spectral scattering coefficient for pure water Morel 1974 (as summarized by Mobley 1994)
}
\author{
Simon Bélanger
}
