% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.cdom.R
\name{spectral.cdom}
\alias{spectral.cdom}
\title{Return CDOM spectra using exponential model}
\usage{
spectral.cdom(wl, a.ref, wl.ref = 440, S = 0.018, K = 0)
}
\description{
Return CDOM spectra using exponential model
}
