% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.nap.R
\name{spectral.nap}
\alias{spectral.nap}
\title{Return non-algal particles (NAP) spectra using exponential model}
\usage{
spectral.nap(wl, a.ref, wl.ref = 440, S = 0.018, K = 0)
}
\description{
Return non-algal particles (NAP) spectra using exponential model
}

