% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.BB9.R
\name{read.BB9}
\alias{read.BB9}
\title{Read s BB-9 file}
\usage{
read.BB9(filen, raw = FALSE)
}
\arguments{
\item{filen}{is a BB-9 file name}

\item{raw}{if a logical parameter indicating whether the file is in raw count}
}
\value{
If calibration already apply, it returns a list with Timer, Beta, BetaP, bbP, bb, wl
}
\description{
Read s BB-9 file
}
\author{
Simon Bélanger
}
